flask_rekognition/flask_rekognition/requirements.txt
flask==3.1.3
flask-login==0.6.3
flask-sqlalchemy==3.1.1
python-dotenv==1.0.1
werkzeug>=3.0.3
boto3==1.34.120
twilio==9.1.0
ultralytics>=8.0.0
opencv-python-headless>=4.9.0
numpy>=2.0.0
Pillow==10.3.0
